Readme updated, added options to convert or not only_audio files

+++ youtube.py
@@ -11,13 +11,14 @@ class Downloader:
     It uses pytube library
     """
 
-    def __init__(self, link, video=None, audio=None, playlist=None, quality='highest'):
+    def __init__(self, link, video=None, audio=None, playlist=None, quality='highest', convert=True):
         self.video = video
         self.audio = audio
         self.playlist = playlist
         self.quality = quality
         self.youtube = Playlist(link) if playlist else YouTube(link)
         self.download_path = './downloads/'
+        self.convert = convert
 
     def _download_video(self, video):
         if self.quality == 'highest':
@@ -36,22 +37,25 @@ class Downloader:
 
         download.download(output_path=self.download_path)
 
-    def _download_audio(self, video):
+    def _download_audio(self, video, convert):
         audio = video.streams.filter(only_audio=True, subtype='mp4').first()
         print('Downloading audio file')
-        path = audio.download()
-        path = Path(path)
-        audio_path = path.name
-        filename = path.stem
-        audiomp3 = AudioSegment.from_file(audio_path)
-        audiomp3.export(self.download_path + filename + '.mp3', format='mp3')
-        os.remove(audio_path)
+        if convert:
+            path = audio.download()
+            path = Path(path)
+            audio_path = path.name
+            filename = path.stem
+            audiomp3 = AudioSegment.from_file(audio_path)
+            audiomp3.export(self.download_path + filename + '.mp3', format='mp3')
+            os.remove(audio_path)
+        else:
+            audio.download(output_path=self.download_path, filename_prefix='only_audio')
 
     def _download_playlist(self):
         for video in self.youtube.videos:
             try:
                 if self.audio:
-                    self._download_audio(video)
+                    self._download_audio(video, self.convert)
                 if self.video:
                     self._download_video(video)
             except exceptions.VideoUnavailable:
@@ -63,7 +67,7 @@ class Downloader:
         else:
             video = self.youtube
             if self.audio:
-                self._download_audio(video)
+                self._download_audio(video, self.convert)
             if self.video:
                 self._download_video(video)
